spss.util.parameterSwitch {SPSSonSparkR}	R Documentation

Automatic switch input value to required format To DO ... 1. switch for scala list? 2. switch for other scala format

Description

Automatic switch input value to required format To DO ... 1. switch for scala list? 2. switch for other scala format

Usage

spss.util.parameterSwitch(parameterMap, parameterName, parameterValue)

[Package SPSSonSparkR version 0.1.0 Index]